moved errorhandler to pulic/js
