+++ public/dashboard.html
@@ -11,7 +11,7 @@
         <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;700&display=swap" rel="stylesheet">
         <link href="https://fonts.googleapis.com/css2?family=DM+Serif+Text:ital@0;1&family=Moderustic:wght@300..800&display=swap" rel="stylesheet">
         <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-        <script src="./js/notificationComponent.js"></script>
+        <script src="./js/notificationComponent.js" type="module"></script>
     </head>
     <body class="dashboard bg-gray-100">
     <nav class="bg-white shadow-md">
@@ -103,7 +103,7 @@
         <p><a href="#">Privacy Policy</a> | <a href="#">Terms of Service</a></p>
     </footer>
 
-    <script>
+    <script <script type="module">
         // Custom error classes
         class ModalError extends Error {
             constructor(message) {
@@ -217,7 +217,7 @@
         }
 
         // Import and use the handleError function from there when needed
-        import { handleError } from './middleware/errorHandler.js';
+        import { handleError } from './js/errorHandler.js';
 
         async function loadNotifications() {
             const notificationComponent = document.getElementById('notification-component');
