+++ server.js
@@ -9,7 +9,7 @@ const FormData = require('form-data');
 const dotenv = require('dotenv');
 const path = require('path');
 const multer = require('multer');
-const errorHandler = require('./middleware/errorHandler');
+const errorHandler = require('./public/js/errorHandler');
 const bodyParser = require('body-parser');
 const authMiddleware = require('./middleware/auth');
 const url = require('url');
